Liftup set interaction. Remove Sets.jsx
